add FunctionTest.cpp
